added validation for edit/add recipe form & register form. also added canDeactivate for edit/add recipe & register forms.

+++ WebClient/src/app/login/register.component.ts
@@ -12,6 +12,8 @@ import { Router } from '@angular/router';
 export class RegisterComponent implements OnInit {
 
   userInfo: IUser;
+  confirmPassword: string;
+  passwordMessage: string;
   message: string;
 
   constructor(
@@ -28,30 +30,38 @@ export class RegisterComponent implements OnInit {
   }
 
   register() {
-    console.log('UserInfo: ' + JSON.stringify(this.userInfo));
-    this.auth.signUp(this.userInfo).subscribe(res => {
-      let tempRes: any;
-      tempRes = res;
+    if (this.userInfo.password !== this.confirmPassword) {
+      this.passwordMessage = 'Passwords do not match.';
+      return;
+    } else {
+      this.passwordMessage = '';
+      console.log('UserInfo: ' + JSON.stringify(this.userInfo));
+      this.auth.signUp(this.userInfo).subscribe(res => {
+        let tempRes: any;
+        tempRes = res;
 
-      // console.log('Res received from node on signup method call: ' + JSON.stringify(res));
+        // console.log('Res received from node on signup method call: ' + JSON.stringify(res));
 
-      if (!res) {
-        console.log('Error creating new user');
-      } else {
-        this.toastr.success('Profile Successfully Created');
-        localStorage.setItem('token', tempRes.token);
-        this.router.navigate(['home']);
-      }
-    },
-    err => {
-      console.log('Err: ' + JSON.stringify(err));
-      this.message = err.error.ErrMessage;
-      this.toastr.error(err.error.ErrMessage);
-    });
+        if (!res) {
+          console.log('Error creating new user');
+        } else {
+          this.toastr.success('Profile Successfully Created');
+          localStorage.setItem('token', tempRes.token);
+          this.router.navigate(['home']);
+        }
+      },
+      err => {
+        console.log('Err: ' + JSON.stringify(err));
+        this.message = err.error.ErrMessage;
+        this.toastr.error(err.error.ErrMessage);
+      });
+    }
+
   }
 
   clearMessage(): void {
     this.message = '';
+    this.passwordMessage = '';
   }
 
 }
